requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
networkx
redis
beautifulsoup4
requests
openai
python-multipart
faiss-cpu 
numpy
python-dotenv
